4.JavaCollections/src/com/javarush/task/task37/task3714/Solution.java
package com.javarush.task.task37.task3714;

import java.io.BufferedReader;
import java.io.IOException;
import java.io.InputStreamReader;
import java.util.ArrayList;
import java.util.LinkedHashMap;
import java.util.Map;

/* 
Древний Рим
*/
public class Solution {
    public static void main(String[] args) throws IOException {
        BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(System.in));
        System.out.println("Input a roman number to be converted to decimal: ");
        String romanString = bufferedReader.readLine();
        System.out.println("Conversion result equals " + romanToInteger(romanString));
    }

    public static int romanToInteger(String s) {

        String[] romeNumbers = {"M", "CM", "D", "CD", "C", "XC", "L", "XL", "X", "IX", "V", "IV", "I"};
        int[] arabicNumbers = {1000, 900, 500, 400, 100, 90, 50, 40, 10, 9, 5, 4, 1};

        s = s.toUpperCase();
        int result = 0;

        int i = 0;
        while (s.length() > 0 && i < romeNumbers.length) {
            if (s.startsWith(romeNumbers[i])) {
                result += arabicNumbers[i];
                s = s.substring(romeNumbers[i].length());
            } else {
                i++;
            }
        }

        return result;
    }
}
